meta-openembedded: iperf: remove depricated package
Now the we have iPerf3, we should drop iPerf. This is the recommendation from the website.
https://sourceforge.net/projects/iperf/

"This is the old, Iperf 2 code base. This version of Iperf is deprecated and we do not recommend it any longer. Please use Iperf3. "

(From meta-openembedded rev: 2b6cad9522582b0816251b5efbbc4601d037ae0e)

Signed-off-by: Armin Kuster <akuster808@gmail.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>