core-image-lsb-sdk: Lower IMAGE_OVERHEAD_FACTOR
We're on the margin of the 4GB size limit for certain image types with
this configuration. This is a temporary workaround to avoid total
failures whilst we figure out better ways to save space. On such a
large image, the standard 1.3 is large and 1.2 is reasonable.

(From OE-core rev: 0116e4814aa55feb0f8944deb98e39b4534bab0c)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>